# Remove reflection usage from `Interpolation` class