This is only intended to remove the reflection from the ctor, both to reduce memory footprint but also to be more AOT-friendly. There appears to be no difference in performance.

Before:

|                   Method |      Mean |     Error |    StdDev | Allocated |
|------------------------- |----------:|----------:|----------:|----------:|
| InterpolateMarginPadding | 45.014 ns | 0.6000 ns | 0.5612 ns |         - |
|  InterpolateFloatGeneric |  8.282 ns | 0.1558 ns | 0.1382 ns |         - |
|         InterpolateFloat |  8.319 ns | 0.1909 ns | 0.2344 ns |         - |

After:

|                   Method |      Mean |     Error |    StdDev | Allocated |
|------------------------- |----------:|----------:|----------:|----------:|
| InterpolateMarginPadding | 45.721 ns | 0.3529 ns | 0.2947 ns |         - |
|  InterpolateFloatGeneric |  7.938 ns | 0.0656 ns | 0.0614 ns |         - |
|         InterpolateFloat |  6.279 ns | 0.1024 ns | 0.0855 ns |         - |