@bdach 2023-04-29
> There appears to be no difference in performance.

I wouldn't expect any benchmark to show any difference in performance given that this PR is touching a static ctor. That said I don't see much a reason *not* to make this change, if only to be a bit friendlier to the JIT (in that it can optimise the unnecessary branches away for each combination of generic type args).

Ran this past game-side tests, no failures. Feeling somewhat confident about this one so I'll just get it in.